Required option in field config accepts lambda
Previously this argument only accepted boolean arguments and didn't
allow you to specify that this could execute code. This code changes
that by allowing lambda functions and symbols (that reference methods)
as arguments.

This has been added so that specifying whether a field is required can
be determined by the openapi_version.