Fields have a concept of allowed
This feature allows fields of a node to be marked as whether they are
allowed to exist or not. This has been added in to support the OpenAPI
3.1 where there are fields that are only valid in OpenAPI 3.1 documents.

A field can take an allowed value of a boolean, a proc or a symbol. It
is expected that, outside of contrived tests, it will never use a plain
boolean as there is no point marking a field as not allowed everywhere.